:recycle: Refactoring for functions
